move videoGalleryUtils.ts to utils folder and renamed to videoGalleryLayoutUtils.ts
